+++ es6/file-type-config.js
@@ -108,6 +108,7 @@ function DocXFileTypeConfig() {
 			"w:sdtContent",
 
 			"w:sectPr",
+			"w:type",
 			"w:headerReference",
 			"w:footerReference",
 		],