+++ es6/tests/base.js
@@ -220,7 +220,6 @@ describe("Inspect module", function () {
 				raw: "#products",
 				lIndex: 15,
 				sectPrCount: 0,
-				sectPrIndex: 0,
 				module: "loop",
 				inverted: false,
 				offset: 0,
@@ -255,7 +254,6 @@ describe("Inspect module", function () {
 						inverted: false,
 						offset: 70,
 						sectPrCount: 0,
-						sectPrIndex: 0,
 						endLindex: 89,
 						lIndex: 89,
 						subparsed: [
@@ -272,7 +270,6 @@ describe("Inspect module", function () {
 								module: "loop",
 								raw: "#proof",
 								sectPrCount: 0,
-								sectPrIndex: 0,
 								inverted: false,
 								offset: 117,
 								endLindex: 133,