+++ es6/tests/fixtures.js
@@ -340,7 +340,6 @@ const fixtures = {
 				module: "loop",
 				inverted: false,
 				sectPrCount: 0,
-				sectPrIndex: 0,
 				subparsed: [{ type: "placeholder", value: "name" }, content(", ")],
 			},
 			endText,
@@ -425,7 +424,6 @@ const fixtures = {
 				value: "users",
 				module: "loop",
 				sectPrCount: 0,
-				sectPrIndex: 0,
 				hasPageBreak: false,
 				hasPageBreakBeginning: false,
 				inverted: false,
@@ -597,7 +595,6 @@ const fixtures = {
 				module: "loop",
 				inverted: false,
 				sectPrCount: 0,
-				sectPrIndex: 0,
 				subparsed: [
 					startParagraph,
 					xmlSpacePreserveTag,